+++ src/extensions/volume-tools/segmentor/labels.ts
@@ -4,7 +4,7 @@
  * @author Tadej Satler <tadej.satler@gmail.com>
  */
 
-import { Volume } from '../../mol-model/volume';
+import { Volume } from '../../../mol-model/volume';
 import { BodyId, BodyInfo, LabelStore } from './types';
 
 const PropertyKey = '__body-labels__';